refactor: extract injectCacheHeaders, use strings.Contains, add YAML tag warnings
Addresses CodeRabbit review feedback:

- Extract injectCacheHeaders() into a named function in main.go, called
  from both the Director closure and tests (no test/production drift)
- Replace custom contains/containsStr with strings.Contains (stdlib)
- Use httptest.NewRequest instead of http.NewRequest in tests
- Add YAML tag collision warning comments on both CLI and server
  VertexAIConfig structs (do NOT yaml:",inline" Anthropic)
